added store param to config, changed output message
added a store parameter to the configuration file, added it to the example json file, and added the checkout link to the sms sent, incase you already have a checkout cart saved on your phone, and want to continue without switching to a computer.

also changed the debug message to specify click and collect instead of store open status.